Moved deduplicateVariableArray to a separate file…
