+++ apps/webapp/app/v3/environmentVariables/environmentVariablesRepository.server.ts
@@ -16,6 +16,7 @@ import {
   type Result,
 } from "./repository";
 import { removeBlacklistedVariables } from "../environmentVariableRules.server";
+import { deduplicateVariableArray } from "../deduplicateVariableArray.server";
 
 function secretKeyProjectPrefix(projectId: string) {
   return `environmentvariable:${projectId}:`;
@@ -840,19 +841,6 @@ export async function resolveVariablesForEnvironment(
     ...projectSecrets,
     ...builtInVariables,
   ]);
-}
-
-/** Later variables override earlier ones */
-export function deduplicateVariableArray(variables: EnvironmentVariable[]) {
-  const result: EnvironmentVariable[] = [];
-  // Process array in reverse order so later variables override earlier ones
-  for (const variable of [...variables].reverse()) {
-    if (!result.some((v) => v.key === variable.key)) {
-      result.push(variable);
-    }
-  }
-  // Reverse back to maintain original order but with later variables taking precedence
-  return result.reverse();
 }
 
 async function resolveOverridableTriggerVariables(
